+++ packages/react-ui/source/component/icons/ArrowUp.tsx
